# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RecruitIQ API — recruitment knowledge management system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /job-openings:
    get:
      operationId: listJobOpenings
      tags: [jobs]
      summary: List all job openings
      responses:
        "200":
          description: List of job openings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobOpening"

    post:
      operationId: createJobOpening
      tags: [jobs]
      summary: Create a new job opening
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOpeningInput"
      responses:
        "201":
          description: Created job opening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOpening"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /job-openings/{id}:
    get:
      operationId: getJobOpening
      tags: [jobs]
      summary: Get job opening by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job opening details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOpening"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateJobOpening
      tags: [jobs]
      summary: Update a job opening
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobOpeningUpdate"
      responses:
        "200":
          description: Updated job opening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobOpening"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteJobOpening
      tags: [jobs]
      summary: Delete a job opening
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /job-openings/{id}/matches:
    get:
      operationId: getJobMatches
      tags: [jobs]
      summary: Get candidates matching a job's role category, sorted by last_verified_at
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matched candidates with their status for this job
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateMatchResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Candidate Job Status ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CandidateWithPreferences:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        resumeText:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        preferences:
          $ref: "#/components/schemas/CandidatePreference"
...
    CandidatePreference:
      type: object
      required: [id, candidateId, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        desiredRole:
          type: ["string", "null"]
        desiredRoleCategory:
          type: ["string", "null"]
        salaryExpectation:
          type: ["string", "null"]
        noticePeriod:
          type: ["string", "null"]
        wfhPreference:
          type: ["string", "null"]
        relocationWillingness:
          type: ["boolean", "null"]
        shiftPreference:
          type: ["string", "null"]
        careerInterests:
          type: ["string", "null"]
        lastVerifiedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    JobOpening:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        roleCategory:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    JobOpeningInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        roleCategory:
          type: string
        requirements:
          type: string
    JobOpeningUpdate:
      type: object
      properties:
        title:
          type: string
        roleCategory:
          type: string
        requirements:
          type: string
    CandidateMatchResult:
      type: object
      required: [candidate]
      properties:
        candidate:
          $ref: "#/components/schemas/CandidateWithPreferences"
        jobStatus:
          $ref: "#/components/schemas/CandidateJobStatus"

    # ── Candidate Job Status ──────────────────────────────────────────────
    CandidateJobStatus:
      type: object
      required: [id, candidateId, jobId, status, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [matched, rejected, open_for_future, interested, contacted]
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
...