# lib/api-zod/src/generated/types/jobOpeningInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */

export interface JobOpeningInput {
  title: string;
  roleCategory?: string;
  requirements?: string;
}